requirements.txt
discord.py>=2.3.2
matplotlib>=3.8.4
python-dotenv>=1.0.1